Refactoring, move subviews to separated files
