+++ fit_airlines_mobile_flutter/lib/models/flight.dart
@@ -1,6 +1,11 @@
 class Flight {
   String name;
   String price;
+  int capacity;
 
-  Flight(this.name, this.price);
+  Flight(
+    this.name,
+    this.price,
+    this.capacity,
+  );
 }
